+++ packages/addon/lua/gelly/api/gx-settings-sync.lua
@@ -8,6 +8,11 @@ local BINARY_MODULE_RELATED_SETTINGS = {
 	"smoothness",
 	"simulation_substeps",
 	"simulation_iterations",
+	"simulation_relaxation",
+	"simulation_collision_distance",
+	"simulation_gravity",
+	"resolution_scale",
+	"glunk_lighting_fix",
 	"max_particles"
 }
 
@@ -16,11 +21,6 @@ function gellyx.settings.updateBinaryModuleSettings(changedConvar)
 	gelly.SetGellySettings({
 		FilterIterations = gellyx.settings.get("smoothness"):GetInt(),
 		EnableGPUSynchronization = true
-	})
-
-	gelly.ConfigureSim({
-		Substeps = gellyx.settings.get("simulation_substeps"):GetInt(),
-		Iterations = gellyx.settings.get("simulation_iterations"):GetInt(),
 	})
 
 	if changedConvar == nil or changedConvar == gellyx.settings.getFullName("max_particles") then
@@ -32,6 +32,35 @@ function gellyx.settings.updateBinaryModuleSettings(changedConvar)
 		hook.Run("GellyRestarted")
 
 		logging.warn("Max particles set to " .. gellyx.settings.get("max_particles"):GetInt() .. "!")
+	end
+
+	if changedConvar == nil or changedConvar == gellyx.settings.getFullName("resolution_scale") then
+		-- Thankfully the sim and renderer are decoupled enough such that we dont
+		-- need to restart Gelly in this case
+		gelly.ChangeResolution(
+			ScrW(),
+			ScrH(),
+			gellyx.settings.get("resolution_scale"):GetFloat()
+		)
+
+		logging.warn("Rendering at %dx%d with a scale of %.2f!", ScrW(), ScrH(),
+			gellyx.settings.get("resolution_scale"):GetFloat())
+	end
+
+	gelly.ConfigureSim({
+		Substeps = gellyx.settings.get("simulation_substeps"):GetInt(),
+		Iterations = gellyx.settings.get("simulation_iterations"):GetInt(),
+		RelaxationFactor = gellyx.settings.get("simulation_relaxation"):GetFloat(),
+		CollisionDistance = gellyx.settings.get("simulation_collision_distance"):GetFloat(),
+		Gravity = gellyx.settings.get("simulation_gravity"):GetFloat()
+	})
+
+	local isLightingFixEnabled = gellyx.settings.get("glunk_lighting_fix"):GetBool()
+
+	if isLightingFixEnabled then
+		RunConsoleCommand("r_worldlights", "0")
+	else
+		RunConsoleCommand("r_worldlights", "4")
 	end
 end