+++ packages/addon/lua/gelly/api/gx-settings.lua
@@ -12,6 +12,11 @@ gellyx.settings.convars = {
 	simulation_rate = settingConvar("simulation_rate", 60, "Controls the simulation rate of the fluid", 20, 144),
 	simulation_substeps = settingConvar("simulation_substeps", 3, "Controls the substeps of the fluid", 1, 100),
 	simulation_iterations = settingConvar("simulation_iterations", 3, "Controls the iterations of the fluid", 1, 100),
+	simulation_relaxation = settingConvar("simulation_relaxation", 1, "Controls the relaxation factor of the solver", 0,
+		1),
+	simulation_collision_distance = settingConvar("simulation_collision_distance", 0.6,
+		"Controls the collision distance of the fluid", 0.01, 1),
+	simulation_gravity = settingConvar("simulation_gravity", -4, "Controls the gravity of the fluid", -100, 100),
 	fluid_radius = settingConvar("fluid_radius", 1, "Controls the radius of the fluid", 1, 5),
 	fluid_viscosity = settingConvar("fluid_viscosity", 1, "Controls the viscosity of the fluid", 0, 100),
 	fluid_cohesion = settingConvar("fluid_cohesion", 1, "Controls the cohesion of the fluid", 0, 1),
@@ -27,6 +32,25 @@ gellyx.settings.convars = {
 	max_particles = settingConvar("max_particles", 100000, "Controls the maximum number of particles (expensive!)", 10000,
 		1500000),
 	sun_visible = settingConvar("sun_visible", 1, "Controls the visibility of the sun (0 = no sun, 1 = sun)", 0, 1),
+
+	gelly_gun_forcefield_strength = settingConvar("gelly_gun_forcefield_strength", -100,
+		"Controls the strength of the forcefield", -1000, 1000),
+	gelly_gun_forcefield_radius = settingConvar("gelly_gun_forcefield_radius", 100,
+		"Controls the radius of the forcefield", 1, 1000),
+	gelly_gun_distance = settingConvar("gelly_gun_distance", 110, "Controls the distance of the emission from the camera",
+		10, 400),
+	gelly_gun_primary_size = settingConvar("gelly_gun_primary_size", 5, "Controls the size of the primary emission", 1,
+		10),
+	gelly_gun_secondary_size = settingConvar("gelly_gun_secondary_size", 5, "Controls the size of the secondary emission",
+		1, 10),
+	gelly_gun_secondary_velocity = settingConvar("gelly_gun_secondary_velocity", 70,
+		"Controls the velocity of the secondary emission", 10, 220),
+	gelly_gun_density = settingConvar("gelly_gun_density", 150, "Controls the density of the emission", 1, 1000),
+
+	resolution_scale = settingConvar("resolution_scale", 0.75,
+		"Controls the rendering scale of the fluid as a percentage of the current resolution.", 0.25, 1.25),
+	glunk_lighting_fix = settingConvar("glunk_lighting_fix", 1, "Controls the lighting fix for Glunk (0 = off, 1 = on)",
+		0, 1),
 }
 
 function gellyx.settings.registerOnChange(name, callback)