+++ packages/gelly/modules/gelly-fluid-renderer/src/v2/renderers/splatting/pipelines/pipeline-info.h
@@ -11,11 +11,11 @@ namespace renderer {
 namespace splatting {
 struct PipelineInfo {
 	const std::shared_ptr<Device> device;
-	const std::shared_ptr<InternalTextures> internalTextures;
-	const std::shared_ptr<OutputTextures> outputTextures;
+	std::shared_ptr<InternalTextures> internalTextures;
+	std::shared_ptr<OutputTextures> outputTextures;
 	const std::shared_ptr<InternalBuffers> internalBuffers;
-	const unsigned int width;
-	const unsigned int height;
+	unsigned int width;
+	unsigned int height;
 };
 }  // namespace splatting
 }  // namespace renderer